Collision masks
With masks it will be able to detect if two objects collide pixel perfectly. However with a cost that this is a more expensive to run.
This is much simpler to use: Only two required parameters.
- Object (E.g: Player, Ai character etc)
- List (Checks if an object in that list has collided with the specified object)

It will return None if no collision has occured. If a collision has occured, it will return the name of the object.